backend/requirements/base.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
django-environ==0.12.0
selenium>=4.15.0